Move OESMILES_OPTIONS to new perses.constants
