+++ perses/constants.py
@@ -1,0 +1,6 @@
+################################################################################
+# CONSTANTS
+################################################################################
+
+from openeye import oechem
+OESMILES_OPTIONS = oechem.OESMILESFlag_DEFAULT | oechem.OESMILESFlag_ISOMERIC | oechem.OESMILESFlag_Hydrogens
